The VTG app now has an upgraded VideoManager. I have fully migrated the videos away from the client and put them on Noel's server so that the app size does not grow too large. This required some research with aSyncTask but I now understand it and the basic concept behind it. It makes things a lot easier :)
I have also fully updated the matrix inconsistancy from before. Before the matrix was incorrect (row, column) with common standards (column, row) and this is now fully updated in the code as well as the videeo references.

The code is fully ready for expansion with more moves. The new naming standard will support the new sets as well as prop videos we decide to add. This was a quick change because all of this is now done in the MatrixID object now... then I had to change all the videos to the correct naming scheme :/